서버 돌아간당

+++ backend/src/main/java/org/iclass/review/entity/StoreReviewStats.java
@@ -1,6 +1,7 @@
 package org.iclass.review.entity;
 
 import jakarta.persistence.*;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -14,25 +15,29 @@ public class StoreReviewStats {
     @Column(name = "rating_count", nullable = false)
     private Integer ratingCount;
 
+    // 0.0 ~ 5.0 범위, 소수점 1자리 예: 4.5
     @Column(name = "avg_rating", nullable = false, precision = 3, scale = 1)
-    private Double avgRating;
+    private BigDecimal avgRating;
 
     // JSON 히스토그램 { "1": n1, "2": n2, ... }
     @Column(name = "rating_histogram", columnDefinition = "JSON")
     private String ratingHistogram;
 
-    @Column(name = "updated_at", nullable = false,
-            columnDefinition = "DATETIME")
+    @Column(name = "updated_at", nullable = false, columnDefinition = "DATETIME")
     private LocalDateTime updatedAt;
 
     public Long getStoreIdx() { return storeIdx; }
     public void setStoreIdx(Long storeIdx) { this.storeIdx = storeIdx; }
+
     public Integer getRatingCount() { return ratingCount; }
     public void setRatingCount(Integer ratingCount) { this.ratingCount = ratingCount; }
-    public Double getAvgRating() { return avgRating; }
-    public void setAvgRating(Double avgRating) { this.avgRating = avgRating; }
+
+    public BigDecimal getAvgRating() { return avgRating; }
+    public void setAvgRating(BigDecimal avgRating) { this.avgRating = avgRating; }
+
     public String getRatingHistogram() { return ratingHistogram; }
     public void setRatingHistogram(String ratingHistogram) { this.ratingHistogram = ratingHistogram; }
+
     public LocalDateTime getUpdatedAt() { return updatedAt; }
     public void setUpdatedAt(LocalDateTime updatedAt) { this.updatedAt = updatedAt; }
 }
